src/requirements.txt
customtkinter==5.2.2
google_api_python_client==2.189.0
google_auth_oauthlib==1.2.4
Pillow==12.1.0
protobuf==6.33.5
pyzipper==0.3.6
